fix: circular dependency in network migration unit test
